:art: rename hijackers to sandbox patchers
