+++ src/sandbox/snapshotSandbox.ts
@@ -7,7 +7,9 @@ import { SandBox } from '../interfaces';
 function iter(obj: object, callbackFn: (prop: any) => void) {
   // eslint-disable-next-line guard-for-in, no-restricted-syntax
   for (const prop in obj) {
-    callbackFn(prop);
+    if (obj.hasOwnProperty(prop)) {
+      callbackFn(prop);
+    }
   }
 }
